[Power] Add support for via-files.
It happens that paf-power can be invoked if more tarmac traces than the
shell or the OS can handle on the command line. The standard way to
solve this problem is to pass the program arguments (or some of them)
thru a file that will be read by the program. This patch implements
support for via-files.

This requires pulling-in a newer revision of tarmac-trace-utilities, and
as a side effect, we get we proper handling of the ELF mapping symbols.